+++ NEWS.md
@@ -61,6 +61,7 @@
 * `any_table()`: Removed c() in examples where not necessary. (24.01.2026)
 * `if.()`: Now outputs a message on how many observations have been removed and how many are left. (27.01.2026)
 * `any_table()`, `combine_into_worbook()`: Adjusted examples. (02.02.2026)
+* `if.()`, `else_if.()`, `else.()`: If used inside a function, these functions should now be able to catch the original variable name passed into the parent function. (10.02.2026)
 
 
 # qol 1.2.0